requirements.txt
customtkinter>=5.2.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
git+https://github.com/openai/CLIP.git
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
